refactor: rename code_generater to code_generator
